@@ -20,7 +20,7 @@ public class ContentResponseConverter extends AbstractResponseConverter {
 	@Override
 	public void accept(Object t, ChatHandlerExecutor u) {
 		if (t instanceof Content) {
-			Message m = t instanceof Message ? (Message) t : Message.of((Content) t);
+			Message m = t instanceof Message m1 ? m1 : Message.of((Content) t);
 			MessageResponse out = new MessageResponse(u.action().getAddressable(), m);
 			rh.accept(out);
 		}
