@@ -200,7 +200,7 @@ public class FreemarkerRendering implements TableRendering<String> {
 	}
 	
 	private String indent(int n) {
-		return "\n"+String.format("%"+n+"s", "");
+		return "\n"+("%" + n + "s").formatted("");
 	}
 
 	@Override
