@@ -103,11 +103,13 @@ public class TestMessageMLParser  {
 	@Test
 	public void testTaggedMessage() throws Exception {
 		EntityJson ej = entityJsonConverter.readValue(
-				"{\"0\":{\"id\":[{\"type\":\"com.symphony.user.userId\",\"value\":\"347583113331315\"}],\"type\":\"com.symphony.user.mention\"},"+
-				 "\"1\":{\"id\":[{\"type\":\"com.symphony.user.userId\",\"value\":\"345315370604167\"}],\"type\":\"com.symphony.user.mention\"},"+
-				 "\"2\":{\"id\":[{\"type\":\"com.symphony.user.userId\",\"value\":\"345315370598706\"}],\"type\":\"com.symphony.user.mention\"},"+
-				 "\"3\":{\"id\":\"bob\",\"type\":\"org.symphonyoss.taxonomy.hashtag\"},"+
-				 "\"4\":{\"id\":[{\"type\":\"org.symphonyoss.fin.security.id.ticker\",\"value\":\"tsla\"}],\"type\":\"org.symphonyoss.fin.security\"}}");
+				"""
+				{"0":{"id":[{"type":"com.symphony.user.userId","value":"347583113331315"}],"type":"com.symphony.user.mention"},\
+				"1":{"id":[{"type":"com.symphony.user.userId","value":"345315370604167"}],"type":"com.symphony.user.mention"},\
+				"2":{"id":[{"type":"com.symphony.user.userId","value":"345315370598706"}],"type":"com.symphony.user.mention"},\
+				"3":{"id":"bob","type":"org.symphonyoss.taxonomy.hashtag"},\
+				"4":{"id":[{"type":"org.symphonyoss.fin.security.id.ticker","value":"tsla"}],"type":"org.symphonyoss.fin.security"}}\
+				""");
 		Message actual = smp.apply(
 			"<div data-format=\"PresentationML\" data-version=\"2.0\" class=\"wysiwyg\"><p> </p><p>/help <span class=\"entity\" data-entity-id=\"3\">#bob</span> <span class=\"entity\" data-entity-id=\"4\">$tsla</span> <span class=\"entity\" data-entity-id=\"0\">@Rob Moffat</span> <span class=\"entity\" data-entity-id=\"1\">@Mark Mainwood</span> <span class=\"entity\" data-entity-id=\"2\">@James Tan</span> </p></div>",
 			ej);
