@@ -39,8 +39,10 @@ public class RSSProperties {
 		this.observationStreamId = observationStream;
 	}
 
-	String failureMessage = "<p>There has been a problem configuring this RSS-Feed.</p>"+
-			"<p>Please talk to the Symphony Administrators for assistance</p>";
+	String failureMessage = """
+			<p>There has been a problem configuring this RSS-Feed.</p>\
+			<p>Please talk to the Symphony Administrators for assistance</p>\
+			""";
 	
 	String observationStreamId = null;		// means don't use
 
